requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.11
idna==3.3
Pillow==9.0.1
pytesseract==0.3.8
requests==2.27.1
urllib3==1.26.8
